+++ packages/palette/src/elements/Cards/LargeCard.shared.tsx
@@ -1,0 +1,8 @@
+import { TagProps } from "./Tag.shared"
+
+export interface LargeCardProps {
+  image: string
+  title: string
+  subtitle?: string
+  tag?: TagProps
+}
