+++ packages/palette/src/elements/Cards/LargeCard.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+import { LargeCardProps } from "./LargeCard.shared"
+
+/**
+ * `Large` is a card with one image one tall image, and text for title and subtitle
+ * at the bottom.
+ */
+export const LargeCard: React.FC<LargeCardProps> = ({}) => {
+  return null
+}
