+++ packages/palette/src/elements/Cards/MediumCard.shared.tsx
@@ -1,0 +1,8 @@
+import { TagProps } from "./Tag.shared"
+
+export interface MediumCardProps {
+  image: string
+  title: string
+  subtitle?: string
+  tag?: TagProps
+}
