+++ packages/palette/src/elements/Cards/MediumCard.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+import { MediumCardProps } from "./MediumCard.shared"
+
+/**
+ * `MediumCard` is a card with one image one tall image, and text for title and subtitle
+ * at the bottom.
+ */
+export const MediumCard: React.FC<MediumCardProps> = ({}) => {
+  return null
+}
