+++ packages/palette/src/elements/Cards/SmallCard.ios.tsx
@@ -1,14 +1,8 @@
 import { Sans, space, Spacer, color, Flex } from "@artsy/palette"
 import React from "react"
 import { Image, View } from "react-native"
-import { Tag, TagProps } from "./Tag.ios"
-
-export interface SmallCardProps {
-  images: string[]
-  title: string
-  subtitle?: string
-  tag?: TagProps
-}
+import { SmallCardProps } from "./SmallCard.shared"
+import { Tag } from "./Tag.ios"
 
 /**
  * `SmallCard` is a card with a layout one square image on the left,
