+++ packages/palette/src/elements/Cards/SmallCard.shared.tsx
@@ -1,0 +1,8 @@
+import { TagProps } from "./Tag.shared"
+
+export interface SmallCardProps {
+  images: string[]
+  title: string
+  subtitle?: string
+  tag?: TagProps
+}
