+++ packages/palette/src/elements/Cards/SmallCard.tsx
@@ -1,0 +1,11 @@
+import React from "react"
+import { SmallCardProps } from "./SmallCard.shared"
+
+/**
+ * `SmallCard` is a card with a layout one square image on the left,
+ * one tall or two square images on the right, and text for title and subtitle
+ * at the bottom.
+ */
+export const SmallCard: React.FC<SmallCardProps> = () => {
+  return null
+}
