+++ packages/palette/src/elements/Cards/Tag.shared.tsx
@@ -1,0 +1,9 @@
+import { StyleProp, ViewStyle } from "react-native"
+
+export interface TagProps {
+  text: string
+  textColor: string
+  color: string
+  borderColor?: string
+  style: StyleProp<ViewStyle>
+}
