+++ packages/palette/src/elements/Cards/Tag.tsx
@@ -1,0 +1,9 @@
+import React from "react"
+import { TagProps } from "./Tag.shared"
+
+/**
+ * `Tag` is used for the Cards, and is controlled by their `tag` prop.
+ */
+export const Tag: React.FC<TagProps> = () => {
+  return null
+}
